+++ Cargo.toml
@@ -1,11 +1,12 @@
 [package]
 name = "interproscan-reader"
-version = "0.1.0"
+version = "0.0.3"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+comfy-table = "5.0.1"
 flate2 = "1.0.22"
 log = "0.4.14"
 pretty_env_logger = "0.4.0"
